requirements.txt
flask==2.0.2
azure-core==1.20.1
azure.cosmos==4.2.0
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
sendgrid==6.9.1
pyyaml==5.4.1